requirements.txt
filetype
pillow
-e .
